Fix_403Error

+++ backend/request/views.py
@@ -11,6 +11,8 @@ from request.models import Cus_req
 def postReq(request):
     if request.method == 'GET':
         userAuth = checkAuth(request)
+        if type(userAuth) == JsonResponse:
+            return userAuth
 
         try:
             targetdata = list(Cus_req.objects.filter(user_uid=userAuth.user_uid) \
@@ -22,6 +24,9 @@ def postReq(request):
 
     if request.method == 'POST':
         userAuth = checkAuth(request)
+        if type(userAuth) == JsonResponse:
+            return userAuth
+
         try:
             inputdata = json.loads(request.body.decode('utf-8'))
             Cus_req.objects.create(
@@ -45,6 +50,9 @@ def postReq(request):
 def fixReq(request, uid):
     if request.method == 'PATCH':
         userAuth = checkAuth(request)
+        if type(userAuth) == JsonResponse:
+            return userAuth
+
         try:
             targetInfo = Cus_req.objects.get(user_uid=userAuth.user_uid, req_uid=uid)
             targetStatus = False
@@ -59,6 +67,9 @@ def fixReq(request, uid):
 
     if request.method == 'DELETE':
         userAuth = checkAuth(request)
+        if type(userAuth) == JsonResponse:
+            return userAuth
+            
         try:
             targetInfo = Cus_req.objects.get(req_uid=uid, user_uid=3)  # userAuth.uid가 들어가야함
             targetInfo.delete()
